+++ server/src/main/java/org/opensearch/authz/package-info.java
@@ -1,0 +1,14 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+/**
+ * Holds classes related to the authorization of actions
+ *
+ * @opensearch.experimental
+ */
+package org.opensearch.authz;